+++ src/Ch02/MyDrop.hs
@@ -16,3 +16,9 @@ myDrop :: Int -> [a] -> [a]
 myDrop n xs = if n <= 0 || null xs
                  then xs
                  else myDrop (n-1) (tail xs)
+
+-- from chapter 3, p. 69
+niceDrop n xs | n <= 0 = xs
+niceDrop _ []          = []
+niceDrop n (_:xs)      = niceDrop (n-1) xs
+