+++ src/Ch03/Lendings.hs
@@ -25,3 +25,11 @@ lend2 amount balance = if amount < reserve * 0.5
                           else Nothing
                             where reserve = 100
                                   newBalance = balance - amount
+
+-- Conditional Evaluation With Guards, p. 69
+lend3 amount balance
+  | amount <= 0            = Nothing
+  | amount > reserve * 0.5 = Nothing
+  | otherwise              = Just newBalance
+  where reserve    = 100
+        newBalance = balance - amount